Fixed some dispatch signatures

+++ src/safeposix/syscalls/fs_calls.rs
@@ -1558,7 +1558,7 @@ impl Cage {
 
     //------------------FTRUNCATE SYSCALL------------------
     
-    pub fn ftruncate_syscall(&self, fd: i32, length: usize) -> i32 {
+    pub fn ftruncate_syscall(&self, fd: i32, length: isize) -> i32 {
         let fdtable = self.filedescriptortable.read().unwrap();
  
         if let Some(wrappedfd) = fdtable.get(&fd) {
@@ -1583,13 +1583,13 @@ impl Cage {
                             let mut fobjtable = FILEOBJECTTABLE.write().unwrap();
                             
                             let mut fileobject = fobjtable.get_mut(&inodenum).unwrap();
-                            let filesize = normalfile_inode_obj.size;
+                            let filesize = normalfile_inode_obj.size as isize;
                             
                             // if length is greater than original filesize,
                             // file is extented with null bytes
                             if filesize < length {
                                 let blankbytecount = length - filesize;
-                                if let Ok(byteswritten) = fileobject.zerofill_at(filesize, blankbytecount as usize) {
+                                if let Ok(byteswritten) = fileobject.zerofill_at(filesize as usize, blankbytecount as usize) {
                                     if byteswritten != blankbytecount as usize {
                                         panic!("zerofill_at() has failed");
                                     }
@@ -1598,7 +1598,7 @@ impl Cage {
                                 }
                             } else { // if length is smaller than original filesize,
                                      // extra data are cut off
-                                fileobject.shrink(length);
+                                fileobject.shrink(length as usize);
                             } 
                             persist_metadata(&mutmetadata);
                         }
@@ -1624,7 +1624,7 @@ impl Cage {
     }
 
     //------------------TRUNCATE SYSCALL------------------
-    pub fn truncate_syscall(&self, path: &str, length: usize) -> i32 {
+    pub fn truncate_syscall(&self, path: &str, length: isize) -> i32 {
         self.ftruncate_syscall(self.open_syscall(path, O_RDWR, S_IRWXA), length)
     }
 
@@ -1675,7 +1675,7 @@ impl Cage {
         
     //------------------GETDENTS SYSCALL------------------
 
-    pub fn getdents_syscall(&self, fd: i32, dirp: *mut u8, bufsize: usize)-> i32 {
+    pub fn getdents_syscall(&self, fd: i32, dirp: *mut u8, bufsize: u32)-> i32 {
         
         let mut vec: Vec<(interface::ClippedDirent, Vec<u8>)> = Vec::new();
 
@@ -1711,7 +1711,7 @@ impl Cage {
                                 vec_filename.push(b'\0'); // make filename null-terminated
                                 
                                 vec_filename.push(DT_UNKNOWN); // push DT_UNKNOWN as d_type (for now)
-                                temp_len = interface::CLIPPED_DIRENT_SIZE + vec_filename.len(); // get length of current filename vector for padding calculation
+                                temp_len = interface::CLIPPED_DIRENT_SIZE + vec_filename.len() as u32; // get length of current filename vector for padding calculation
                                 
                                 // pad filename vector to the next highest 8 byte boundary
                                 for _ in 0..(temp_len + 7) / 8 * 8 - temp_len {
@@ -1719,7 +1719,7 @@ impl Cage {
                                 }
                                 
                                 // the fixed dirent size and length of filename vector add up to total size
-                                curr_size = interface::CLIPPED_DIRENT_SIZE + vec_filename.len(); 
+                                curr_size = interface::CLIPPED_DIRENT_SIZE + vec_filename.len() as u32;
                                 
                                 bufcount += curr_size; // increment bufcount
